refactor: Move JWT and video upload logic to utils, remove unused video controllers
