--- utils/jwt.js
@@ -1,20 +1,0 @@
-const jwt = require("jsonwebtoken");
-
-const generateToken = (userId) => {
-  return jwt.sign({ userId }, process.env.JWT_SECRET, {
-    expiresIn: process.env.JWT_EXPIRES_IN || "7d",
-  });
-};
-
-const verifyToken = (token) => {
-  try {
-    return jwt.verify(token, process.env.JWT_SECRET);
-  } catch (error) {
-    return null;
-  }
-};
-
-module.exports = {
-  generateToken,
-  verifyToken,
-};
